requirements2.txt
actionlib==1.11.13
angles==1.9.11
attrdict==2.0.0
backports.weakref==1.0.post1
base-local-planner==1.14.4
bleach==1.5.0
bondpy==1.8.3
camera-calibration==1.12.23
camera-calibration-parsers==1.11.13
capabilities==0.2.0
catkin==0.7.14
catkin-pkg==0.4.2
click==6.7
controller-manager==0.13.3
controller-manager-msgs==0.13.3
create-driver==2.3.1
create-node==2.3.1
cv-bridge==1.12.8
diagnostic-analysis==1.9.3
diagnostic-common-diagnostics==1.9.3
diagnostic-updater==1.9.3
docutils==0.14
dynamic-reconfigure==1.5.50
empy==3.3.2
eventlet==0.19.0
Flask==0.11.1
funcsigs==1.0.2
gazebo-plugins==2.5.18
gazebo-ros==2.5.18
gencpp==0.6.0
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.11
gennodejs==2.0.1
genpy==0.6.7
greenlet==0.4.13
h5py==2.6.0
html5lib==0.9999999
image-geometry==1.12.8
interactive-markers==1.11.4
itsdangerous==0.24
Jinja2==2.10
joint-trajectory-action-tools==0.0.10
Keras==2.0.8
laser-geometry==1.6.4
Markdown==2.6.11
MarkupSafe==1.0
message-filters==1.12.14
mock==2.0.0
moveit-commander==0.9.15
moveit-ros-planning-interface==0.9.15
moveit-ros-visualization==0.9.15
numpy==1.13.1
openni2-launch==0.4.2
pbr==4.0.3
Pillow==2.2.1
pluginlib==1.11.3
pr2-controller-manager==1.8.17
protobuf==3.5.2.post1
pyparsing==2.2.0
python-dateutil==2.7.3
python-engineio==2.1.1
python-qt-binding==0.3.4
python-socketio==1.6.1
PyYAML==3.12
qt-dotgraph==0.3.11
qt-gui==0.3.11
qt-gui-cpp==0.3.11
qt-gui-py-common==0.3.11
resource-retriever==1.12.4
rocon-app-manager==0.9.1
rocon-app-utilities==0.9.1
rocon-apps==0.9.1
rocon-console==0.3.2
rocon-ebnf==0.3.2
rocon-gateway==0.8.1
rocon-gateway-utils==0.8.1
rocon-hub==0.8.1
rocon-hub-client==0.8.1
rocon-interactions==0.3.2
rocon-master-info==0.3.2
rocon-python-comms==0.3.2
rocon-python-redis==0.3.2
rocon-python-utils==0.3.2
rocon-python-wifi==0.3.2
rocon-semantic-version==0.3.2
rocon-uri==0.3.2
rosbag==1.12.14
rosboost-cfg==1.14.4
rosclean==1.14.4
roscreate==1.14.4
rosgraph==1.12.14
roslaunch==1.12.14
roslib==1.14.4
roslint==0.11.0
roslz4==1.12.14
rosmake==1.14.4
rosmaster==1.12.14
rosmsg==1.12.14
rosnode==1.12.14
rosparam==1.12.14
rospkg==1.1.4
rospy==1.12.14
rosservice==1.12.14
rostest==1.12.14
rostopic==1.12.14
rosunit==1.14.4
roswtf==1.12.14
rqt-action==0.4.9
rqt-bag==0.4.12
rqt-bag-plugins==0.4.12
rqt-console==0.4.8
rqt-dep==0.4.9
rqt-graph==0.4.9
rqt-gui==0.5.0
rqt-gui-py==0.5.0
rqt-image-view==0.4.13
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.8
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.0
rqt-py-console==0.4.8
rqt-reconfigure==0.4.10
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.8
rqt-robot-steering==0.5.9
rqt-runtime-monitor==0.5.7
rqt-rviz==0.5.10
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.10
rqt-web==0.4.8
rviz==1.12.16
scipy==0.19.1
sensor-msgs==1.12.7
six==1.11.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.3
srdfdom==0.4.2
tensorflow-gpu==1.3.0
tensorflow-tensorboard==0.1.8
tf==1.11.9
tf-conversions==1.11.9
tf2-geometry-msgs==0.5.20
tf2-kdl==0.5.20
tf2-py==0.5.20
tf2-ros==0.5.20
topic-tools==1.12.14
unique-id==1.0.5
urdfdom-py==0.3.3
Werkzeug==0.14.1
xacro==1.11.3
